ci(*) add Docker build and push workflow
Add a workflow to build and push Docker images for 2.x based on the new
2.x-only file structure.

Add a static file containing the latest version. This must be updated
manually. If the pushed tag matches the latest version, Docker
automation will add a "latest" tag when pushing image tags.

Add logic to check the latest version file and tag images latest if
their tag matches the version in the latest_version file.