feat(gateway) ignore listener past
Remove any awareness of Listener history when calculating status other
than the attachedRoutes figure from Status. If any two Listeners result
in a conflict condition, mark all Listeners with that port or hostname
conflicted. Adding a new Listener in conflict with an existing Listener
will break the existing Listener and detach it.